add node
Yes, it's a prominent project. I'm one of the most prolific contributors
the last few years, I'm on the project's Technical Steering Committee,
and the web folks have explicitly expressed skepticism about doing this.
I'm just grabbing it to redirect to the main Node.js site to prevent
anyone from using it to redirect to evilsite.example.com. I'll happily
hand it over to the nodejs GitHub org if they want it (and, in fact, I'd
almost assuredly be the person doing the transfer as I'm one of the
owner's of that org).