api.js.org
Add `api` t the list.

Not sure if this is allowed, but i plan to make a api network with a bunch of apis built by the open source community, currently i have a rps(rock, paper, scissors) api.
It can be viewed at https://rps://rps.gamertike.com (i plan to move to api.js.org if possible). This is prove i have some content on my page.