Added 'rx' CNAME entry pointing to iaseth.github.io/readmix
Readmix is a markdown/HTML/PDF? generator that allows you to write JavaScript in markdown.
The readmix cli is available at https://www.npmjs.com/package/readmix.
I am using 'rx' instead of 'readmix' for the subdomain because 'rx' is the file extension used by readmix templates, and it is shorter then the fullname of the repo.